+++ R/add.R
@@ -16,17 +16,54 @@
 #' This function is really meant to be used in a pipeline as a quick shortcut to add the
 #' linked status column
 #'
-#' @param .data a data frame (or tibble) with a studyid column
+#' @param .data a data frame (or tibble) or arrow Dataset with a studyid column
+#' @param studyid_col the name of the study id column. Defaults to "studyid"
 #'
 #' @return the same data frame plus a `linked_status` column
 #'
 #' @export
+add_linked_status_col <- function(.data, studyid_col = "studyid"){
+
+  UseMethod("add_linked_status_col")
+
+}
+
+#' @export
+add_linked_status_col.default <- function(.data, studyid_col = "studyid"){
+  stop("No add_linked_status_col method for an object of class ", class(.data), call. = FALSE)
+}
 
 
-add_linked_status_col <- function(.data) {
-  if (!"studyid" %in% names(.data)) stop("No studyid column", call. = FALSE)
+#' @export
+add_linked_status_col.data.frame <- function(.data, studyid_col = "studyid") {
+  dplyr::mutate(
+    .data, linked_status = ifelse(
+      grepl("^s", !!rlang::sym(studyid_col)), "linked", "unlinked"
+      )
+    )
+}
 
-  .data$linked_status <- ifelse(grepl("^s", .data$studyid), "linked", "unlinked")
+#' @export
+add_linked_status_col.Dataset <- add_linked_status_col.data.frame
 
-  .data
+
+#' @export
+add_linked_status_col.arrow_dplyr_query <- add_linked_status_col.Dataset
+
+#' Wrapper to filter out unlinked studyids from either a data.frame or a Dataset
+#'
+#' This is a convenience function which automatically filter for only "linked" studyids
+#' as defined by `add_linked_status()`.
+#'
+#' @inheritParams add_linked_status_col
+#' @inheritDotParams add_linked_status_col
+#'
+#' @export
+#'
+#' @examples
+#'
+#' filter_linked(data.frame(studyid = c("sxxxx", "uxxxx")))
+filter_linked <- function(.data, ...) {
+    d <- add_linked_status_col(.data, ...)
+    dplyr::filter(d, linked_status == "linked")
 }