+++ tests/testthat/test-read-dat.R
@@ -5,13 +5,30 @@ test_that("reads in file",{
                       data_dict = dict)
 
   expect_is(raw, "data.frame")
+
+  raw2 <- read_dat(dat_path_csv, data_format = "csv")
+  expect_is(raw2, "data.frame")
+
+  expect_equivalent(raw, raw2)
 })
+
 
 test_that("reads in two files via vector", {
   files_vec <- c(dat_path, dat_path2)
   two_files <- read_dat(
     data_path = files_vec,
     data_dict = dict,
+    id = "file"
+  )
+  expect_is(two_files, "tbl_df")
+  expect_identical(files_vec, unique(two_files$file))
+})
+
+test_that("reads in two csv files via vector", {
+  files_vec <- c(dat_path_csv, dat_path_csv2)
+  two_files <- read_dat(
+    data_path = files_vec,
+    data_format = "csv",
     id = "file"
   )
   expect_is(two_files, "tbl_df")