+++ NEWS.md
@@ -1,3 +1,5 @@
+# dipr (development version)
+
 # dipr 1.0.0
 
 ## Breaking Change