+++ NEWS.md
@@ -5,6 +5,7 @@
   2. Removing files that can't be imported - these are listed in the `_ocwaignore` file in the root of the repo
   3. Cleaning `README.md` to comment out references to images and links that won't be available in the SRE.
   4. Committing the changes from 2 and 3 to the new branch and pushing that to GitHub. This branch can then be used as the basis for an import into OCWA.
+* Added `dipr_create_targets_project` which will create a thin package-like targets folder structure. 
 
 # dipr 1.0.0